trying to see if i am handling interrupts properly

+++ src/interconnect.rs
@@ -77,8 +77,8 @@ impl Interconnect {
         }
     }
 
-    pub fn emu_cycles(&mut self, cpu_cycles: i32) {
-        let n: i32 = cpu_cycles * 4;
+    pub fn emu_cycles(&mut self, cpu_cycles: u64) {
+        let n: u64 = cpu_cycles * 4;
         for _ in 0..n {
             self.timer.internal_ticks = self.timer.internal_ticks.wrapping_add(1);
             self.timer_tick();
